// src/app/features/rentals/models/model/model.ts
import { Entity } from "../entity";

export interface Model extends Entity{
  id: number;
  brandId: number;
  fuelId: number;
  transmissionId: number;
  name: string;
  dailyPrice: number;
  imageUrl: string;
}
export interface ModelList {
  id: number;
  brandName: string;
  fuelName: string;
  transmissionName: string;
  name: string;
  dailyPrice: number;
  imageUrl: string;
}